Driving controls

+++ src/org/usfirst/frc/team79/robot/commands/StartIntake.java
@@ -9,7 +9,7 @@ import edu.wpi.first.wpilibj.command.Command;
  */
 public class StartIntake extends Command {
 	
-	private int invert;
+	private double invert;
 
 	/**
 	 * Toggles the intake
@@ -24,7 +24,7 @@ public class StartIntake extends Command {
     }
     
     public void execute() {
-    		Robot.intake.intake.set(-1.0*invert);
+    	Robot.intake.intake.set(-1.0*invert);
     }
 
     public boolean isFinished() {
